Add get parking lot status API and get all vehicles in a parking lot API. Add flyway and migration scripts

+++ src/main/java/com/royeen/smartpark/business/usecase/GetParkingLotOccupancyCommandImpl.java
@@ -12,7 +12,7 @@ public class GetParkingLotOccupancyCommandImpl implements GetParkingLotOccupancy
     private final ParkingTransactionRepository repository;
 
     @Override
-    public long execute(String lotId) {
+    public int execute(String lotId) {
         return repository.countByParkingLotIdAndParkingStatus(lotId, ParkingStatus.CHECKED_IN);
     }
 }
